+++ src/__spec_helper__/index.js
@@ -1,6 +1,8 @@
 import Enzyme from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import { setup } from "./mock-match-media";
+import setupResizeObserverMock from "./mock-resize-observer";
 
+setupResizeObserverMock();
 setup();
 Enzyme.configure({ adapter: new Adapter() });