+++ src/components/select/multi-select/index.d.ts
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Option from "../option";
+import { OptionProps } from "../option";
 
 export interface MultiSelectProps {
   /** Boolean to toggle where SelectList is rendered in relation to the Select Input */
@@ -33,7 +33,7 @@ export interface MultiSelectProps {
   /** The default selected value(s), when the component is operating in uncontrolled mode */
   defaultValue?: string[] | object[];
   /** Child components (such as Option) for the SelectList */
-  children: Array<typeof Option>;
+  children: Array<React.ReactElement<OptionProps>> | React.ReactElement<OptionProps>;
   /** If true the Component opens on focus */
   openOnFocus?: boolean;
   /** A custom message to be displayed when any option does not match the filter text */