+++ CHANGELOG.md
@@ -1,3 +1,92 @@
+### [66.12.1](https://github.com/Sage/carbon/compare/v66.12.0...v66.12.1) (2021-02-22)
+
+
+### Bug Fixes
+
+* **select:** incorrect typescript children types ([b436fea](https://github.com/Sage/carbon/commit/b436feadad57d92a5b6ae640b88ae33dc9ba9f99))
+
+## [66.12.0](https://github.com/Sage/carbon/compare/v66.11.1...v66.12.0) (2021-02-22)
+
+
+### Features
+
+* **radio-button:** allow use of all styled system margin props ([4698606](https://github.com/Sage/carbon/commit/46986061d1bddef119d95c2d60eb83e0d9f1e1d5))
+
+### [66.11.1](https://github.com/Sage/carbon/compare/v66.11.0...v66.11.1) (2021-02-21)
+
+
+### Bug Fixes
+
+* **date-range:** allow empty values for start and end dates ([01eb5bc](https://github.com/Sage/carbon/commit/01eb5bc1cb58fb4a084fc2b9e85009960621d20c)), closes [#3501](https://github.com/Sage/carbon/issues/3501)
+
+## [66.11.0](https://github.com/Sage/carbon/compare/v66.10.2...v66.11.0) (2021-02-19)
+
+
+### Features
+
+* **flat-table:** add expandable rows option ([9701362](https://github.com/Sage/carbon/commit/9701362ba12e1a4e26881981848d1cb500473c92))
+
+### [66.10.2](https://github.com/Sage/carbon/compare/v66.10.1...v66.10.2) (2021-02-19)
+
+
+### Bug Fixes
+
+* **date:** set format based on translation ([024f338](https://github.com/Sage/carbon/commit/024f338ce7c1df99589ab9154c8aaf8c0071fdb0))
+
+### [66.10.1](https://github.com/Sage/carbon/compare/v66.10.0...v66.10.1) (2021-02-19)
+
+
+### Bug Fixes
+
+* **focus-trap:** handle radio button focus ([ce3cc9b](https://github.com/Sage/carbon/commit/ce3cc9bfd0f456a663380af5f6cdcd035cb4e930)), closes [#3623](https://github.com/Sage/carbon/issues/3623)
+
+## [66.10.0](https://github.com/Sage/carbon/compare/v66.9.0...v66.10.0) (2021-02-19)
+
+
+### Features
+
+* **tab:** allows Tab to be a link ([96e1597](https://github.com/Sage/carbon/commit/96e1597faeb3912f811729ae972b0d50eb6a4e6d))
+
+## [66.9.0](https://github.com/Sage/carbon/compare/v66.8.0...v66.9.0) (2021-02-18)
+
+
+### Features
+
+* **select:** add popperjs positioning mechanism ([9bd1a74](https://github.com/Sage/carbon/commit/9bd1a74962903c0b3c8f9c2455227701e210afec))
+
+
+### Bug Fixes
+
+* **multi-select:** fix closing multi-select when deleting pills ([80c95de](https://github.com/Sage/carbon/commit/80c95defa941128402a159c97cb19e8bf5d9f53a))
+
+## [66.8.0](https://github.com/Sage/carbon/compare/v66.7.0...v66.8.0) (2021-02-18)
+
+
+### Features
+
+* **text-editor:** add rows prop to support custom min  height ([966d405](https://github.com/Sage/carbon/commit/966d405be3e4b9aff3b28d01633a2108b7996c63)), closes [#3632](https://github.com/Sage/carbon/issues/3632)
+
+## [66.7.0](https://github.com/Sage/carbon/compare/v66.6.1...v66.7.0) (2021-02-18)
+
+
+### Features
+
+* **help:** add support for tooltip background and font color overrides ([c169979](https://github.com/Sage/carbon/commit/c16997917ce6184ec67d31eb0510bf90e723a0a2))
+* **icon:** add support for tooltip background and font color overrides ([d4f5db9](https://github.com/Sage/carbon/commit/d4f5db934c27ea44b688f35a169ddb79c461e9be))
+* **tooltip:** add support for background and font color overrides ([f607f77](https://github.com/Sage/carbon/commit/f607f77c973c71c79198b20d3609c68782f5da24))
+
+
+### Bug Fixes
+
+* **tooltip:** ensure long text strings wrap instead of overflow ([de44e2b](https://github.com/Sage/carbon/commit/de44e2b34bc3ce8f347b71fe552edc1ed2f9d981))
+
+### [66.6.1](https://github.com/Sage/carbon/compare/v66.6.0...v66.6.1) (2021-02-17)
+
+
+### Bug Fixes
+
+* **date:** change onBlur event parameters ([16d186d](https://github.com/Sage/carbon/commit/16d186d09794863c7d042fa8b3c7ae20262b0bf7)), closes [#3132](https://github.com/Sage/carbon/issues/3132)
+
 ## [66.6.0](https://github.com/Sage/carbon/compare/v66.5.0...v66.6.0) (2021-02-17)